Move leaderboard files to their separate folders
